Separated auth into Auth Service.
