+++ client/src/data/datasources/pammicc-api.js
@@ -35,44 +35,3 @@ export async function newProject(newProject) {
   }
 }
 
-export async function login(user) {
-  const response = await fetch(`${server_url}/login`, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(user),
-  })
-  .catch(error => {
-    throw error;
-  })
-
-  if (response.status == 401)  {
-    throw new Error("Please provide a valid email and password combination.")
-  }
-
-  if (!response.ok) {
-    throw new Error("An error ocurred login user.")
-  }
-}
-
-export async function register(newUser) {
-  const response = await fetch(`${server_url}/register`, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(newUser),
-  })
-  .catch(error => {
-    throw error;
-  })
-
-  if (response.status == 409) {
-    throw new Error("This email is already registered.")
-  }
-
-  if (!response.ok) {
-    throw new Error("An error ocurred registering user.")
-  }
-}
